Improved NextTurnCalculator and EndPhase
To follow up with the previous change, I created overloads for the NextTurnCalculator and StartEndPhase to start the End Phase under other conditions. This should improve and fix the EndTurnButton in a more efficient and multiplayer-friendly way, but it has not been tested and may need a couple of patches to work out if not correct.