Better CombatCalculation & Team Logic
Some minor improvements to performance and readability for the CombatCalculation, as well as improved team designation (such as fixing everybody in FFA being on the same team) and team logic (bots now no longer attack teammates). Overhauls to heart and club logic will be rolled out over the next couple of commits.