Champion Selection & Ability Implementation
This commit finally polishes the sandbox UI pre-game, allowing the player to choose their own champion! As for the bot, currently they pick randomly from the pool of available champions.

This commit also shows an example of a simple ability being used! As such, a new champion has been added to the game: the Regime Captain. The ability added to the Regime Captain is known as Quick Assist, an ability that activates on the champion's Beginning Phase, granting that champion an additional card per champion in the game in their faction (excluding themselves).

Speaking of factions, another enum has been added that introduces faction-sorting to the game.

Other bug fixes that completely break the game have also been fixed.

The next commit should aim to improve upon the hotswappable ability system design, and also work towards bringing gamemodes to the pre-game UI.