Ability Panel
This commit adds an ability panel accessible by right-clicking the Champion you are looking at. Currently it is only cosmetically working, but functionality should be implemented over the course of the next commits. I also want to make sure that you could only use abilities that corresponds to your player, so that the player cannot use OTHER champions' abilities.