Redesigned Ability System
I took the time to move all of the ability checking to it's own script, as well as implementing the functionality of the ability panel in conjunction with the ability system, and I'm glad to report that it works quite well! Important things have been changed, such the ability panels being instantiated at game start, and then being teleported on screen when it is needed (as well as teleported off-screen when it is no longer needed.) The next step in order is to provide more feedback to the user when an ability is activated to make it feel more of a mechanic than gimmick.