Hoplomachus Champion, New Abilities, and Bug Fixes
A new champion, Hoplomachus, has been added to the game! Currently, because of his superior number of abilities, is considered the superior champion in the game.

This also means new abilities to the game.
Hoplite Tradition: An exclusive passive ability that allows the user to have a 50% chance to heal for 20 when another champion is dealt a Queen or higher. (This excludes the starting hand and club trading.)
Hoplite Shield: An exclusive defense bonus that allows the user to have a 20% chance to cut incoming damage by half. This chance increases to 50% if the damage is lethal.

This update also addresses numerous bugs and loops that were caused by the migration of a universal card playing system, but I was too stupid and ignorant to find on the first update. This means that this is the first update to be fully playable since commits from a week ago ;(.
Apart from the biggest bugs, some values have also been tweaked to make the bot more aggressive and prone to attacking, as I found in beta testing that the bots rarely attacked the player and/or made risks that might've saved/cost their game. Hopefully tweaking these values make the game more dynamic.

In the next upcoming commits, I am going to focus on creating multiple difficulty modes, so that beginners can play the game without trouble whilst veterans (like me) can still destroy the hardest difficulty. I would also like to make a tutorial scene soon, but that's probably going to take a while, so stay tuned for that.