Quick Heal & Bojutsu
Added two new abilities to the game, as well as more functionality for abilities in general.

Quick Heal: A generic passive ability that has a 50% chance of healing the champion for double the amount. (It sounds quite broken on paper, only time/beta-testing will tell.)

Bojutsu: An exclusive defense bonus that gives an attack penalty towards attacks with a Jack or higher. (How this works: if the attack disadvantage is 1, then the value of the card will calculated as the card value - 1. If someone attacked with a Queen [value of 12 in the game], then it will be calculated as 12 - 1 = 11, as if it were a Jack.)