Difficulty Levels
A full functioning difficulty level has been implemented to the game! Previously, the game was set to it's maximum difficulty permanently, and there was no way to change it (it was even hardcoded and somehow hacking and changing the enum would not work.) Now, there are four difficulty levels: Noob, Novice, Warrior, and Champion. Keep in mind that the maximum difficulty has also been amped up further to make it more challenging. Here's a breakdown of the difficulties:
Noob - The bot does not know how to play cards correctly, the dealing of cards is rigged for the player, and is slow as a snail. A recommended difficulty for beginners who have no experience with strategy/card games in general. However, the prize reward will be on average 1/5 of what it would be playing on a more modest difficulty.
Novice - This gamemode is similar to Noob, but the bot has a slight bit more sense and the cards are fully random from this gamemode upwards. This gamemode could be great for practicing and/or newcomers to this particular game, but have a bit of experience with strategy games. Keep in mind that the gold reward is still halved.
Warrior - This gamemode is the default gamemode. For those who know the ins-and-outs of the game, and is pretty good at every mechanic. The bot knows basic strategies and can properly play cards. This gamemode is great for sharpening your skills with new abilities against tougher opponents that know what they're doing. Full gold reward included.
Champion - This gamemode is specially designed for veterans of the strategy game genre and is designed to challenge those who dare. The bot knows common and sophisticated scenarios and can hold his own against even the upper-range of pros. Since this difficulty is a bit of a challenge to beat, the gold reward has a 20% multiplier for the risk-takers who dare to challenge themselves.

Have fun playing, and please share any game-breaking bugs with the game at this point! The game itself haven't been tested after this update, so game-breaking bugs could go unnoticed for another couple of commits before I eventually find out, so your help is appreciated!