Better GameEndPanel & Bug Fixes
Since the end scene looked kind of out of place and badly made, and by badly made I mean absolutely fucking atrocious, I decided to completely remove it and just integrate it with the game end area in the same scene. This also allows me to properly fix all the issues with the game end bonuses not working properly, and also introduces a way to finally see how many gold you currently have, as well as a cool animation showing the gaining of gold in order to give more verbose and feedback to the user, and make the user feel accomplishment and progression.

Obviously, gold doesn't do anything at the moment, but I'm already currently working on a way for gold to be able to be used, as well as integrating this with the champion selection. Stay tuned!