LeanTween Workarounds
Supposedly referencing the Transform of a GameObject will produce awkward and stupid results, and this has plagued me for the better part of the last week. This commit transfers all LeanTween.scale and LeanTween.move function calls to reference the RectTransform instead of the Transform itself, as UI elements work better with references to the RectTransform as referencing the Transform itself may break what the RectTransform reads. This supposedly also fixed every issue with the UI elements not tweening correctly.

yes I will finally work on a store now