Shop Concept & More Deprecation
This commit finally adds a working shop concept into the game. You still have access to all of the champions, but this will allow you to test out the shop and how it would look in the end! I am still planning on adding a confirm dialog for when someone purchases an item, as well as a way to stop someone from buying certain items if they will be in debt. Overall, this commit will be good!

Some other things to note: I'm currently testing sorting champions by their alphabetical name.