Error Tooltips
A working error tooltip has been implemented into the game. This has no drastic difference compared to a normal tooltip, with the exception of a red color to signify a warning.
Typically, warning tooltips should disappear on its own, whereas a normal tooltip only disappears when a certain criteria is met, such as hiding itself whenever the mouse is no longer over a certain object.