Champion Serialization Improvements
Previously, champions were stored as instance IDs when bought from the store. This meant that, at least in the Unity editor, that reloading or relaunching the Unity editor would remove all of the champions from your possession and throw a NullReferenceException or some sort of argument error. The DataManager now saves and loads from a list of IDs that is assigned to each champion, which stays the same for throughout every launch and will not change due to an instance change.