Champion Slots & Spawning Overhaul
Champion slots have been added to the game, and currently they will indicate how/where champions will spawn at runtime. The introduction of champion slots and the improved spawning allows for dynamic spawning and easy management of where a champion is on screen at any given time. This also means that the overlapping of champions if spawned incorrectly is no longer the case, and that spawning minions or champions mid-game is now possible and easier than ever.

I have also noticed an issue regarding Arya and her reportedly "going missing in the editor after a while from the sandbox champion selector". This might have to do with the way champions are serialized, and I will look into that and attempt to fix that in the next commit.