Unsynced DataManager Fix
This commit fixes the issue that has to do with the DataManager not serializing and loading correctly. Previously, there were two instances of the DataManager running due to how the singleton worked, but now it has been fixed and all instances of the DataManager have been fixed. This also means that the bug that removes certain champions from the player's possession has been fixed.