GambleButton, Dealing & GetAttackingCard Fixes
The GambleButton used to cause a soft-lock if pressed out of turn. This has been patched.
The GambleButton would also previously not disappear even if a card was selected. This has been hopefully patched.
After the last commit, the card that was played would wait for the function it caused to finish first before it was discarded, since all of the `Deal()` coroutines are now set to `yield return`. This has been patched.
In Warrior & Champion Difficulty, if GetAttackingCard returned a null, a NullReferenceException would occur since there was no code to handle the null returned by GetAttackingCard. This has also been patched.
Another similar issue with CombatCalculation was that if the champion had reviewed the choices and found that they didn't want to attack with their current configuration, a NullReferenceException would occur and the game would soft-lock, do to the fact that the game does not know what to do after the bot decides not to attack. This has been patched.
A small fix has also been issued to correctly show the percentage of health left when calculating the reward at the end of a match.
Finally, the delay time before showing the tooltip when hovering over a card has been increased from 0.5 seconds to 1 second, to make accidental pop-ups happen less when scrolling through cards.