Confirm Dialog Improvements
This commit improved visuals of the confirm dialog, as well as makes the confirm dialog more flexible.

Added a static instance to the current confirm dialog that is present. This is automatically set to `null` when Hide() is called, and is only set when a ConfirmDialog is created by the static `CreateNew` function. This also allows for accessing the current confirm dialog from anonymous functions, and allows for cleaner code.

It is now advised to use lambda expressions or delegates to declare the action of both buttons, as this allows for the action to be passed parameters.

Other than that, the font for the buttons have been switched out for Roboto-Black.