Ability Feed Improvement
This commit improves the AbilityFeed for ChampionControllers, allowing for multiple feed entries to be shown at once, as well as grouping the same feed entries by name and showing that they are grouped by showing how many times a feed entry was repeated. This was done by creating a grid layout group on an empty GameObject with a script that controls the spawning and displaying of new ability feeds.

Other than the obvious, the Champion Template has also received reference improvements, and no longer get their references by the index of child. Instead, all of the components of the Champion Template are now explicitly referenced in the editor itself.