Minion Support
Added a new field in the ChampionController script that holds the currentOwner of the ChampionController. If this field is null, it is considered a standard champion. If this field is set to another ChampionController, it is considered a minion of that ChampionController.

A function to spawn a ChampionController as a minion of another ChampionController has also been added.

Finally, ChampionControllers considered as a minion will not be counted when checking for the end-of-game criteria. This means that, for example, if the owner is dead but the owner owned two minions, they will not count towards the final player count when calculating the end-of-game criteria. This means that if your original opponent is the last player standing, then they win in FFA even though there are two minions left.