AudioController Features
Added several new features to the AudioController.

The AudioController can now play clips from other AudioSources via an overload of the Play() function. If loop is turned off using the overload with the `AudioClip` parameter.

Currently, the background music of the Sandbox scene and the glass-break effect in the Main Menu use this method, but future changes will be introduced in the later future.