World Space Canvas + EZCameraShake
This commit moves all of the primary canvases to the `World Space`. This means that the canvases are now independent of the camera, and that the camera's movement will not influence the canvas' position. This also means that camera shake is finally doable on the UI canvas. Furthermore, this shouldn't affect anything that has to do with gameplay or cause any bugs since the camera itself isn't actually moving or being moved anywhere other than by the camera shake.

EZCameraShake has also been re-imported to the project.

Currently, the camera will shake when the game focuses on the main menu, when someone takes damage, and when a match ends and the winner avatar is displayed.

This also fixes some bugs, notably to do with Diamond Logic (since the bot logic was being applied to the player as well, leading to some unfortunate errors and bugs.)