Combat Gambling Authority Fix
This commit fixes an issue that would prevent the champion from discarding their attacking or defending card, due to the attacking or defending card being gambled and thus technically not counting as their hand's card.
This has been fixed by giving an exception to the discard authority of the hand. Discarding cards now ignores if the card is part of their hand or "technically the hand's owner's" if it is the owner's attacking or defending card.

Oh and I also messed with some material stuff nothing was applied though