Critical Hand Fixes & More External Clip Porting
This commit fixes a critical mistake on my behalf that was made last commit. In short, discarding cards have now been fixed, and attacking and defending cards will now ignore the authority check correctly.

As for the external audio clip porting, the Hoplite Shield and Smite sound effects have been moved to the ability ScriptableObject itself, and is no longer tied to the AudioController prefab. We intend to do this for every other ability in the future, as well as every other unique sound effect too.