Code Improvements & ParticleSystem Work
This commit optimizes and adds to the GameEnd and GameEndCheck functions in hopes of optimizing the code and adding support for future gamemodes. In addition, this commit also adds some work done to the particle systems of the ChampionControllers.

This commit also addresses countless other issues. The previous commits to fix the issue with not being able to discard gambled cards due to the authority has been finally fixed.
Other less important changes include the removal of no-longer-necessary Debug.Logs, and slight changes to scenes.